feat(46): lot of fix
