+++ services/horizon/interfaces/http/plugins/accelerator.ts
@@ -1,0 +1,4 @@
+import { ZodAccelerator } from "@duplojs/core";
+import { operatorContentSchema } from "@vendors/types-advanced-query";
+
+ZodAccelerator.build(operatorContentSchema);
