+++ services/marine-snow/interfaces/providers/scienceDatabase/pubmed/index.ts
@@ -68,7 +68,7 @@ export class PubMedAPI {
 			.catch((error: unknown) => error as Error);
 	}
 
-	public static getArticle(pubmedId: string) {
+	public static getArticle(pubmedId: string[]) {
 		return retry(
 			() => this.httpClient.get(
 				"/entrez/eutils/efetch.fcgi",
